// kitchen-planner-fe-main/src/app/api/api-kms-planner-masterdata/model/order.ts
/**
 * KMS - Kitchen Management System API
 *
 * Contact: anas-ur-rasheed.abid@tu-ilmenau.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DeliveryAddress } from './deliveryAddress';
import { BillingInfo } from './billingInfo';


export interface Order { 
    _id?: string;
    meals?: Array<object>;
    total?: number;
    subTotal?: number;
    tax?: number;
    deliveryFee?: number;
    discount?: number;
    status?: Order.StatusEnum;
    billingInfo?: BillingInfo;
    deliveryAddress?: DeliveryAddress;
    specialInstructions?: string;
    createdAt?: string;
}
export namespace Order {
    export const StatusEnum = {
        Pending: 'pending',
        Confirmed: 'confirmed',
        Preparing: 'preparing',
        Ready: 'ready',
        Delivered: 'delivered',
        Cancelled: 'cancelled'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}